feat: Adds support for deleting AMIs
Adds support for deleting AMIs and associated snapshots. Also improves
the logic of the delete_vpc function to check for any attached security
groups and deletes them before attempting to delete the VPC.
Additionally, this commit adds error handling to the delete_ec2_instance
function to safely handle cases in which the EC2 instance has been
terminated and does not have an ARN.
The new function get_other_resources should be used to get resources
that cannot be retrieved with the resource-groups client. Any functions
for retrieving those resources should be placed in the get_other_ids
file.