fix: Improves pagination logic for retieval
Improves login in get_resources_by_tag to include MaxResults to 500 for
both the initial get_resources call, and also in the event there is a
pagination token. In the very unlikely event that this is ever needed,
this will help to reduce API calls.

Additionally, a short waiter is placed between regions for the region
loop.