refactor: ASG retrieval placed in get_other_ids
Formerly, ASGs were retrieved inside of the get_resources_by_tag
function. Now they are retrieved from get_autoscaling_groups, which is
inside of get_other_ids, and called by get_other_resources. This has
been tested and works.

Additionally makes a minor fix from the last commit in which MaxResults
were set to 100. Upon testing, it was discovered that 50 was the limit,
so that is to 50 as well.