feat: Adds create_dynamodb_table_backup
Adds new function create_dynamodb_table_backup which is called by
delete_dynamodb_table. The function is called if the table is not empty
and prompts users if they would like to create a backup.

Also adds a new text formatting function custom_prompt, which allows
input of any string, not just "y" or "n".